Rearrange modules, tweak Context semantics, support OpenGL
